update constant definitions
