+++ src/core/api_gate_mt.hpp
@@ -58,7 +58,7 @@ class APIGateMultiThread : public APIGateBase, public ControllerDelegate {
   void controller_on_reply(Controller& sm, std::unique_ptr<api::Reply> reply) override;
   void controller_on_require_invoke(Controller& sm, unsigned int msec) override;
 
-  void logger_on_output(Logger& logger, LogLevel::Type level, const std::string& message) override;
+  void logger_on_output(Logger& logger, LogLevel level, const std::string& message) override;
 
   void loop_event();
   void loop_controller();
